gen-test: don't expose diff + attributions
